update for long long when cout the total

+++ src/Headers/BookManager.cpp
@@ -280,7 +280,7 @@ void BookManager::buy(std::string str, std::string quant) {
                 return;
             } else {
                 item.Inventory -= quantity;
-                int total = quantity * item.Price;
+                long long total = (long long)quantity * (long long)item.Price;
                 std::cout << std::fixed << std::setprecision(2);
                 std::cout << (double)total / 100.00 << "\n";
                 LM.addinfo((double)total / 100.00);
